backend/requirements.txt
pymupdf
tqdm
qdrant-client
sentence-transformers
python-dotenv
elasticsearch
groq
onnxruntime>=1.17.0
transformers>=4.40.0
torch>=2.1.0
numpy>=1.26.0
pytest>=7.4.0
fastapi>=0.111.0
python-multipart>=0.0.9
uvicorn>=0.29.0
psycopg2-binary>=2.9.9